add component ContactDetailsSection
